functions/vectorizer/requirements.txt
pydantic
asyncpg
pgvector
python-dotenv
ruff
mypy
pdfplumber
python-docx
asyncio
langchain_text_splitters 
openai